fix: types location
