+++ src/components/button/Button.tsx
@@ -1,6 +1,15 @@
 import React from 'react';
 import style from './Button.module.scss';
-import { ButtonProps } from '../types/interface'
+import { MouseEventHandler, ReactNode } from "react";
+
+export interface ButtonProps {
+  children: ReactNode;
+  onClick?: MouseEventHandler<HTMLButtonElement>;
+  variant?: string;
+  size?: string;
+  type?: string;
+  className?: string;
+}
 
 export const Button: React.FC<ButtonProps> = ({ children, onClick, variant = 'primary', size = 'medium' }) => {
   return (
