+++ src/components/links/HeaderLinks.tsx
@@ -1,6 +1,16 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { HeaderLinksProps } from "../types/interface";
+
+export interface HeaderLink {
+  name: string;
+  href: string;
+  authOnly?: boolean;
+}
+
+export interface HeaderLinksProps {
+  links: HeaderLink[];
+  user?: any;
+}
 
 const HeaderLinks: React.FC<HeaderLinksProps> = ({ links, user }) => {
   return (
