+++ src/components/loadingBar/LoadingBar.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import style from './LoadingBar.module.scss';
-import { LoadingBarProps } from '../types/interface';
+
+export interface LoadingBarProps {
+  variant?: string; 
+}
 
 const LoadingBar: React.FC<LoadingBarProps> = ({variant = 'loaderSpinner' }) => {
   return <div className={`${style[variant]}`}></div>;
