+++ src/pages/logInPage/LogInPage.tsx
@@ -4,7 +4,11 @@ import { auth } from "../../components/api/firebaseConfig";
 import LoadingBar from "../../components/loadingBar/LoadingBar";
 import { Button } from "../../components/button/Button";
 import "./loginPage.css";
-import { LogInPageProps } from "../../components/types/interface";
+
+export interface LogInPageProps {
+  user: User | null;
+  isAuthLoading: boolean;
+}
 
 const LogInPage: React.FC<LogInPageProps> = ({ user, isAuthLoading }) => {
   const [email, setEmail] = useState("");
